refactor : 폴더 data 상위 컴포넌트로 이동
